Parallelizing (w/ GENERATORS) Perceptual Hashing Step
Updating the perceptual hashing part of the app to parallelize the process of fetching and phashing the videos. For each video hash that doesn't already have a perceptual hash in the db, we spin off a new process to handle fetching the video from the Client API and then phashing it.

Because the Sqlitedict object can't be pickled (and thus can't be passed to new processes), we can't have the parallel jobs do the checking for if a hash is already in the DB or not. So we have to check all the video hashes for cached values before we can start perceptual hashing. I didn't put a progress bar on that part, but we might want one for larger libraries.